refactor: card types
