+++ projects/ion/src/lib/card/card.component.ts
@@ -1,5 +1,3 @@
-import { IonChipProps } from './../chip/chip.component';
-import { ComponentType } from '@angular/cdk/overlay';
 import {
   AfterViewInit,
   ChangeDetectorRef,
@@ -12,51 +10,8 @@ import {
   ViewChild,
   ViewContainerRef,
 } from '@angular/core';
-import { IconType } from '../icon/icon.component';
 import { SafeAny } from '../utils/safe-any';
-import { DropdownItem } from '../dropdown/dropdown.component';
-
-type Pick<T, K extends keyof T> = {
-  [Key in K]: T[Key];
-};
-
-type ButtonType = 'primary' | 'secondary' | 'ghost';
-
-type ButtonBase = {
-  type: ButtonType;
-  nameAction: string;
-  label: string;
-  icon?: string;
-  circular?: boolean;
-};
-
-type Header = {
-  title: string;
-  buttons?: ButtonBase[];
-  chips?: IonChipProps[];
-  icon?: IconType;
-};
-
-type FooterButton = {
-  [keys in ButtonType]?: Pick<ButtonBase, 'label' | 'circular' | 'icon'>;
-};
-
-type Foote = {
-  buttons?: FooterButton;
-  body?: ComponentType<SafeAny>;
-};
-
-export type CardEvent = {
-  buttonAction?: ButtonType;
-  chipSelected?: { chip: IonChipProps; index: number };
-  selectedFromChipDropdown?: DropdownItem[];
-};
-
-export interface IonCard {
-  header: Header;
-  body: ComponentType<SafeAny>;
-  footer?: Foote;
-}
+import { IonCard, CardEvent } from '../core/types/card';
 
 @Component({
   selector: 'ion-card',
