+++ projects/ion/src/lib/core/types/index.ts
@@ -2,3 +2,4 @@ export * from './direction';
 export * from './info-badge';
 export * from './status';
 export * from './tooltip';
+export * from './card';
